# docs/openapi.yaml
openapi: 3.1.0
info:
  title: BizEngine API
  version: 1.0.0
  description: Business engine for managing organizations, products, orders, warehouses, HR, finance, logistics, and Russian government integrations.
servers:
  - url: http://localhost:8080
    description: Local development
  - url: https://api.bizengine.app
    description: Production
...
paths:
...
  /api/v1/auth/register:
    post:
      tags: [Auth]
      summary: Register a new user
      operationId: register
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [email, password, name]
              properties:
                email: { type: string, format: email }
                password: { type: string, minLength: 8 }
                name: { type: string }
      responses:
        '200':
          description: Registration successful, session cookies set
  /api/v1/auth/login:
    post:
      tags: [Auth]
      summary: Log in
      operationId: login
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [email, password]
              properties:
                email: { type: string }
                password: { type: string }
      responses:
        '200':
          description: Login successful
...
  /api/v1/organizations/{orgID}/catalog/products:
    get:
      tags: [Catalog]
      summary: List products
      operationId: listProducts
      parameters:
        - $ref: '#/components/parameters/orgID'
      responses:
        '200':
          description: Product list
    post:
      tags: [Catalog]
      summary: Create product
      operationId: createProduct
      parameters:
        - $ref: '#/components/parameters/orgID'
      responses:
        '201':
          description: Product created
  /api/v1/organizations/{orgID}/orders:
    get:
      tags: [Orders]
      summary: List orders
      operationId: listOrders
      parameters:
        - $ref: '#/components/parameters/orgID'
      responses:
        '200':
          description: Order list
    post:
      tags: [Orders]
      summary: Create order
      operationId: createOrder
      parameters:
        - $ref: '#/components/parameters/orgID'
      responses:
        '201':
          description: Order created
...
  /api/v1/organizations/{orgID}/warehouse/{warehouseID}/stock:
    get:
      tags: [Warehouse]
      summary: List stock levels
      operationId: listStock
      parameters:
        - $ref: '#/components/parameters/orgID'
        - name: warehouseID
          in: path
          required: true
          schema: { type: string, format: uuid }
      responses:
        '200':
          description: Stock levels
...
  /api/v1/organizations/{orgID}/hr/employees:
    get:
      tags: [HR]
      summary: List employees
      operationId: listEmployees
      parameters:
        - $ref: '#/components/parameters/orgID'
      responses:
        '200':
          description: Employee list
    post:
      tags: [HR]
      summary: Hire employee
      operationId: hireEmployee
      parameters:
        - $ref: '#/components/parameters/orgID'
      responses:
        '201':
          description: Employee hired
  /api/v1/organizations/{orgID}/finance/accounts:
    get:
      tags: [Finance]
      summary: List accounts
      operationId: listAccounts
      parameters:
        - $ref: '#/components/parameters/orgID'
      responses:
        '200':
          description: Account list
  /api/v1/organizations/{orgID}/finance/transactions:
    post:
      tags: [Finance]
      summary: Create transaction
      operationId: createTransaction
      parameters:
        - $ref: '#/components/parameters/orgID'
      responses:
        '201':
          description: Transaction created
  /api/v1/organizations/{orgID}/logistics/routes:
    get:
      tags: [Logistics]
      summary: List delivery routes
      operationId: listRoutes
      parameters:
        - $ref: '#/components/parameters/orgID'
      responses:
        '200':
          description: Route list
...
  /api/v1/organizations/{orgID}/export:
    post:
      tags: [Export]
      summary: Export data
      operationId: exportData
      parameters:
        - $ref: '#/components/parameters/orgID'
      requestBody:
        content:
          application/json:
            schema:
              type: object
              required: [format, type]
              properties:
                format: { type: string, enum: [csv, pdf] }
                type: { type: string }
      responses:
        '200':
          description: File download
  /api/v1/organizations/{orgID}/integrations/fiscal/receipt:
    post:
      tags: [Integrations]
      summary: Send fiscal receipt (ATOL)
      operationId: sendReceipt
      parameters:
        - $ref: '#/components/parameters/orgID'
      responses:
        '201':
          description: Receipt sent
  /api/v1/organizations/{orgID}/integrations/edo/documents:
    post:
      tags: [Integrations]
      summary: Send EDO document (Diadoc)
      operationId: sendEDODocument
      parameters:
        - $ref: '#/components/parameters/orgID'
      responses:
        '201':
          description: Document sent
  /api/v1/organizations/{orgID}/integrations/marking/verify:
    post:
      tags: [Integrations]
      summary: Verify marking code (Chestniy Znak)
      operationId: verifyMarking
      parameters:
        - $ref: '#/components/parameters/orgID'
      responses:
        '200':
          description: Verification result
  /api/v1/organizations/{orgID}/integrations/bank/import:
    post:
      tags: [Integrations]
      summary: Import bank statement (1C format)
      operationId: importBankStatement
      parameters:
        - $ref: '#/components/parameters/orgID'
      responses:
        '200':
          description: Statement imported
...
  /api/v1/organizations/{orgID}/webhooks:
    get:
      tags: [Webhooks]
      summary: List webhooks
      operationId: listWebhooks
      parameters:
        - $ref: '#/components/parameters/orgID'
      responses:
        '200':
          description: Webhook list
    post:
      tags: [Webhooks]
      summary: Create webhook
      operationId: createWebhook
      parameters:
        - $ref: '#/components/parameters/orgID'
      requestBody:
        content:
          application/json:
            schema:
              type: object
              required: [url, events]
              properties:
                url: { type: string, format: uri }
                events: { type: array, items: { type: string } }
                secret: { type: string }
      responses:
        '201':
          description: Webhook created
...
  /api/v1/organizations/{orgID}/white-label:
    get:
      tags: [Banking]
      summary: Get white-label banking configuration
      operationId: getWhiteLabel
      parameters:
        - $ref: '#/components/parameters/orgID'
      responses:
        '200':
          description: White-label config
          content:
            application/json:
              schema:
                type: object
                properties:
                  ok: { type: boolean }
                  data:
                    type: object
                    properties:
                      bank_name: { type: string }
                      bank_code: { type: string }
                      connected: { type: boolean }
                      features: { type: array, items: { type: string } }
  /api/v1/organizations/{orgID}/payments/initiate:
    post:
      tags: [Banking]
      summary: Initiate a payment
      operationId: initiatePayment
      parameters:
        - $ref: '#/components/parameters/orgID'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [data]
              properties:
                data:
                  type: object
                  required: [recipient_account, recipient_bik, recipient_name, recipient_inn, amount, currency, purpose]
                  properties:
                    recipient_account: { type: string }
                    recipient_bik: { type: string }
                    recipient_name: { type: string }
                    recipient_inn: { type: string }
                    amount: { type: integer, description: Amount in minor units (kopecks) }
                    currency: { type: string, enum: [RUB] }
                    purpose: { type: string }
      responses:
        '201':
          description: Payment initiated
          content:
            application/json:
              schema:
                type: object
                properties:
                  ok: { type: boolean }
                  data:
                    type: object
                    properties:
                      payment_id: { type: string, format: uuid }
                      status: { type: string, enum: [pending, processing, completed, failed] }
                      created_at: { type: string, format: date-time }
  /api/v1/organizations/{orgID}/payments/{paymentID}/status:
    get:
      tags: [Banking]
      summary: Check payment status
      operationId: getPaymentStatus
      parameters:
        - $ref: '#/components/parameters/orgID'
        - name: paymentID
          in: path
          required: true
          schema: { type: string, format: uuid }
      responses:
        '200':
          description: Payment status
          content:
            application/json:
              schema:
                type: object
                properties:
                  ok: { type: boolean }
                  data:
                    type: object
                    properties:
                      payment_id: { type: string, format: uuid }
                      status: { type: string, enum: [pending, processing, completed, failed] }
                      amount: { type: integer }
                      currency: { type: string }
                      created_at: { type: string, format: date-time }
                      updated_at: { type: string, format: date-time }
                      error_message: { type: string }
  /api/v1/organizations/{orgID}/finance/bank/auto-reconcile:
    post:
      tags: [Banking]
      summary: Auto-reconcile bank transactions
      operationId: autoReconcile
      parameters:
        - $ref: '#/components/parameters/orgID'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [data]
              properties:
                data:
                  type: object
                  required: [bank_account_id, date_from, date_to]
                  properties:
                    bank_account_id: { type: string, format: uuid }
                    date_from: { type: string, format: date }
                    date_to: { type: string, format: date }
      responses:
        '200':
          description: Reconciliation result
          content:
            application/json:
              schema:
                type: object
                properties:
                  ok: { type: boolean }
                  data:
                    type: object
                    properties:
                      reconciliation_id: { type: string, format: uuid }
                      matched: { type: integer }
                      unmatched: { type: integer }
                      auto_matched: { type: integer }
                      manual_review: { type: integer }
                      status: { type: string }
  /api/v1/organizations/{orgID}/finance/bank/reconciliations:
    get:
      tags: [Banking]
      summary: List reconciliation runs
      operationId: listReconciliations
      parameters:
        - $ref: '#/components/parameters/orgID'
      responses:
        '200':
          description: Reconciliation list
          content:
            application/json:
              schema:
                type: object
                properties:
                  ok: { type: boolean }
                  data:
                    type: array
                    items:
                      type: object
                      properties:
                        id: { type: string, format: uuid }
                        bank_account_id: { type: string, format: uuid }
                        date_from: { type: string, format: date }
                        date_to: { type: string, format: date }
                        matched: { type: integer }
                        unmatched: { type: integer }
                        status: { type: string }
                        created_at: { type: string, format: date-time }
...
  /api/v1/auth/bank/{bankCode}/login:
    get:
      tags: [Banking]
      summary: Redirect to bank OAuth login
      operationId: bankOAuthLogin
      parameters:
        - name: bankCode
          in: path
          required: true
          schema: { type: string }
      responses:
        '302':
          description: Redirect to bank authorization page
  /api/v1/auth/bank/{bankCode}/callback:
    get:
      tags: [Banking]
      summary: Bank OAuth callback
      operationId: bankOAuthCallback
      parameters:
        - name: bankCode
          in: path
          required: true
          schema: { type: string }
        - name: code
          in: query
          schema: { type: string }
        - name: state
          in: query
          schema: { type: string }
      responses:
        '200':
          description: Bank account linked
          content:
            application/json:
              schema:
                type: object
                properties:
                  ok: { type: boolean }
                  data:
                    type: object
                    properties:
                      bank_code: { type: string }
                      account_linked: { type: boolean }
...
components:
  parameters:
    orgID:
      name: orgID
      in: path
      required: true
      schema:
        type: string
        format: uuid
...